+++ MainMenu/GraphAction.java
@@ -10,7 +10,7 @@ public class GraphAction
 {
     private String color;
     private String playerid;
-    private Double nodeid;
+    private Integer nodeid;
     private String action;
 
     /**
@@ -29,11 +29,11 @@ public class GraphAction
         this.color = color;
     }
     
-    public Double getNodeId(){
+    public Integer getNodeId(){
         return nodeid;
     }
     
-    public void setNodeId(Double nodeid){
+    public void setNodeId(Integer nodeid){
         this.nodeid = nodeid;
     }
